fix useBookmark
