+++ web/client-v2/src/features/articles/components/Card/ArticleCardWrapper/useArticleBookmark.ts
@@ -1,8 +1,7 @@
-import { FragmentOf, readFragment } from "gql.tada";
-import { useCallback, useState } from "react";
+import { useMutation } from "@apollo/client";
+import { FragmentOf, graphql, readFragment } from "gql.tada";
+import { useCallback } from "react";
 
-import { createArticleBookmarkMutation } from "@/features/articles/actions/actCreateArticleBookmarkMutation";
-import { deleteArticleBookmarkMutation } from "@/features/articles/actions/actDeleteArticleBookmarkMutation";
 import { logoutToLoginPage } from "@/features/auth/actions/auth";
 import { getUser } from "@/features/auth/actions/user";
 
@@ -10,67 +9,92 @@ import { useStatusToast } from "@/hooks/useStatusToast";
 
 import { ArticleCardItemFragment } from "../ArticleCardItem";
 
+const CreateArticleBookmarkMutation = graphql(`
+  mutation CreateBookmarkMutation($input: CreateBookmarkInput!) {
+    createBookmark(createBookmarkInput: $input) {
+      id
+    }
+  }
+`);
+
+const DeleteArticleBookmarkMutation = graphql(`
+  mutation DeleteBookmarkMutation($input: DeleteBookmarkInput!) {
+    deleteBookmark(deleteBookmarkInput: $input)
+  }
+`);
+
 export const useArticleBookmark = (
   data: FragmentOf<typeof ArticleCardItemFragment>
 ) => {
   const fragment = readFragment(ArticleCardItemFragment, data);
-  const [bookmarkId, setBookmarkId] = useState<string | null>(
-    fragment.bookmarkId
-  );
   const { successToast, failToast } = useStatusToast();
 
-  const handleAddBookmark = useCallback(
-    async (articleId: string) => {
-      if (!fragment.platform) return;
-      const user = await getUser();
-      if (!user) {
-        failToast({
-          description: "Fail: Please login to bookmark this article",
-        });
-        await logoutToLoginPage();
-        return;
-      }
-
-      const { data, error } = await createArticleBookmarkMutation({
-        articleId,
-        userId: user.id,
-        platformId: fragment.platform?.id,
-        title: fragment.title,
-        description: fragment.description,
-        articleUrl: fragment.articleUrl,
-        thumbnailUrl: fragment.thumbnailUrl,
-        publishedAt: fragment.publishedAt,
-        platformName: fragment.platform?.name,
-        platformUrl: fragment.platform?.siteUrl,
-        platformFaviconUrl: fragment.platform?.faviconUrl,
-        isEng: fragment.isEng,
-        isRead: false,
-      });
-
-      if (error) {
-        if (error.length > 0) {
-          failToast({
-            description: error[0].message,
-          });
-          return;
-        }
-        failToast({
-          description: "Fail: Something went wrong",
-        });
-        return;
-      }
-
-      if (data?.createBookmark?.id) {
-        successToast({
-          description: "Add bookmark",
-        });
-        setBookmarkId(data.createBookmark.id);
-      }
-    },
-    [fragment, successToast, failToast]
+  const [createArticleBookmarkMutation] = useMutation(
+    CreateArticleBookmarkMutation
   );
 
-  const handleRemoveBookmark = useCallback(
+  const [deleteArticleBookmarkMutation] = useMutation(
+    DeleteArticleBookmarkMutation
+  );
+
+  const handleCreateBookmark = useCallback(async () => {
+    const user = await getUser();
+    if (!user) {
+      failToast({
+        description: "Fail: Please login to bookmark this article",
+      });
+      await logoutToLoginPage();
+      return;
+    }
+
+    const { errors } = await createArticleBookmarkMutation({
+      variables: {
+        input: {
+          articleId: fragment.id,
+          userId: user.id,
+          platformId: fragment.platform?.id,
+          title: fragment.title,
+          description: fragment.description,
+          articleUrl: fragment.articleUrl,
+          thumbnailUrl: fragment.thumbnailUrl,
+          publishedAt: fragment.publishedAt,
+          platformName: fragment.platform?.name || "",
+          platformUrl: fragment.platform?.siteUrl || "",
+          platformFaviconUrl: fragment.platform?.faviconUrl || "",
+          isEng: fragment.isEng,
+          isRead: false,
+        },
+      },
+      update: (cache, { data }) => {
+        cache.modify({
+          id: cache.identify(fragment),
+          fields: {
+            isBookmarked: () => true,
+            bookmarkId: () => data?.createBookmark.id,
+          },
+        });
+      },
+    });
+
+    if (errors) {
+      if (errors.length > 0) {
+        failToast({
+          description: errors[0].message,
+        });
+        return;
+      }
+      failToast({
+        description: "Fail: Something went wrong",
+      });
+      return;
+    }
+
+    successToast({
+      description: `Add bookmark title 【 ${fragment.title} 】`,
+    });
+  }, [createArticleBookmarkMutation, fragment, successToast, failToast]);
+
+  const handleDeleteBookmark = useCallback(
     async (bookmarkId: string) => {
       const user = await getUser();
       if (!user) {
@@ -81,15 +105,28 @@ export const useArticleBookmark = (
         return;
       }
 
-      const { data, error } = await deleteArticleBookmarkMutation({
-        bookmarkId,
-        userId: user.id,
+      const { errors } = await deleteArticleBookmarkMutation({
+        variables: {
+          input: {
+            bookmarkId,
+            userId: user.id,
+          },
+        },
+        update: (cache) => {
+          cache.modify({
+            id: cache.identify(fragment),
+            fields: {
+              isBookmarked: () => false,
+              bookmarkId: () => null,
+            },
+          });
+        },
       });
 
-      if (error) {
-        if (error.length > 0) {
+      if (errors) {
+        if (errors.length > 0) {
           failToast({
-            description: error[0].message,
+            description: errors[0].message,
           });
           return;
         }
@@ -99,15 +136,12 @@ export const useArticleBookmark = (
         return;
       }
 
-      if (data?.deleteBookmark) {
-        successToast({
-          description: "Remove bookmark",
-        });
-        setBookmarkId(null);
-      }
+      successToast({
+        description: `Remove bookmark title 【 ${fragment.title} 】`,
+      });
     },
-    [successToast, failToast]
+    [deleteArticleBookmarkMutation, fragment, failToast, successToast]
   );
 
-  return { bookmarkId, handleAddBookmark, handleRemoveBookmark };
+  return { handleCreateBookmark, handleDeleteBookmark };
 };
